chore: rename (jwks-)handler as per convention
